refactor: update SelectTabs component
