+++ src/templates/Explore/NewCommunityPoolsTable/index.tsx
@@ -26,9 +26,7 @@ import { Pool_OrderBy } from '@/gql/generated/kassandraApi'
 import SkeletonLoading from '@/components/SkeletonLoading'
 import GradientLabel from '@/components/Labels/GradientLabel'
 import Label from '@/components/Labels/Label'
-import { handleGetAPR } from '@/components/StakeCard/utils'
-import { networks } from '@/constants/tokenAddresses'
-import { ZeroAddress } from 'ethers'
+
 import { useGetAprData } from '@/hooks/query/useGetAprData'
 
 type UnderlyingAssets = {
